feat(settings): introduce a settings-section registry for the Application page
Convert the Application settings page from a fixed JSX sequence into a host
that renders sections from a registry, gated by a small host-owned
SettingsContext. Existing controls are extracted into three first-party,
natively-rendered React sections (General, Conversation Titles, Git), each of
which owns its own persistence (section-owned save).

This is the first, smallest slice of a broader contribution/plugin model: the
{ id, page, order, when } envelope is the stable contract, and Component is the
only producer expected to change in later phases. Backend-specific (and, later,
plugin-contributed) sections can be added by registration + a `when` predicate
instead of by adding `backend.kind`/`app_mode` conditionals to the shared page.

- src/settings/registry.ts: register/get sections, idempotent by id, order+id
  sort, throwing `when` treated as hidden.
- src/settings/use-settings-context.ts: derives SettingsContext (backend kind).
- app-settings/sections/{general,conversation-titles,git}-section.tsx: the
  extracted sections with per-section save buttons.
- src/settings/register-app-settings-sections.ts: registers the built-ins.
- routes/app-settings.tsx: renders getSettingsSections() for the page.

Tests: registry unit tests; app-settings tests updated for per-section submit
buttons. Field test IDs are unchanged.

Refs #16596
